fix(LdesDiskWriter): Internal Quad structure cannot handle relative URIs so use a base IRI during processing
Relative IRIs are allowed in Turtle, but when the N3 Parser parses the Turtle to their internal general Quad structure, no relative URIs are allowed and the parser tries to interpret them as full URIs. Without passing a base URI, the deepness of the relative URIs is getting lost, so we fix this by passing a custom base IRI with enough path levels so we can restore the relative URIs after processing the content with the N3 Parser and Writer.